Move payments to billing directory
